+++ WinGame/objects/Land.h
@@ -62,8 +62,8 @@ private:
 	void setRotate(float degree) override;
 	GVector2 getAnchorPoint() override;
 	
-	void setZIndex(int z) override;
-	int getZIndex() override;
+	void setZIndex(float z) override;
+	float getZIndex() override;
 #pragma endregion
 };
